requirements.txt
# Google Drive API
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# LinkedIn API / HTTP Requests
requests==2.31.0

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# Timezone handling
pytz==2024.1

# Testing
pytest==8.0.0
